Making rolling command arguments optional.
These values can mostly be inferred if not specified. But for TotalChunks
there is a default value that is set in the config file. Migrated rolling
options from the BroStaticCfg to their own RollingStaticCfg struct.

Additionally, non-rolling databases can be converted to rolling databases
by importing new data into them.